抽奖工具dev/requirements.txt
Pillow==9.1.1
PyMySQL==1.0.2
ttkbootstrap==1.7.2
xlrd==1.2.0
